diabetes_rag/requirements.txt
python==3.10.13

torch==2.2.1
transformers==4.38.2
accelerate==0.27.2

sentence-transformers==2.6.1
faiss-gpu==1.7.4

langchain==0.1.16
langgraph==0.0.40

streamlit==1.32.2

numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1.post1

biopython==1.83

# 🔥 PERFORMANCE
vllm==0.4.0